backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-dotenv>=1.0.1

# Adicionado para análise de mídia
python-magic-bin>=0.4.14
google-generativeai>=0.5.4

# Adicionado para upload de arquivos via formulário
python-multipart>=0.0.9